pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "monzo-mcp"
version = "0.1.0"
description = "MCP server for read-only access to Monzo banking data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Anastasia Krause" }
]
keywords = ["monzo", "mcp", "claude", "banking", "api", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/monzo-mcp"
Documentation = "https://github.com/yourusername/monzo-mcp#readme"
Repository = "https://github.com/yourusername/monzo-mcp.git"
Issues = "https://github.com/yourusername/monzo-mcp/issues"

[project.scripts]
monzo-mcp = "monzo_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/monzo_mcp"]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["monzo_mcp"]

# Mypy configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = []

[[tool.mypy.overrides]]
module = ["mcp.*", "httpx.*"]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=monzo_mcp --cov-report=term-missing"

# Coverage configuration
[tool.coverage.run]
source = ["src/monzo_mcp"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
